dll_0031_minimap: s16 return for Camera_GetViewportYOffset + u16 cast on getButtonsJustPressed
Target sign-extends Camera_GetViewportYOffset's return (extsh) and masks
getButtonsJustPressed to u16 (clrlwi r3,r3,16) at the call site. Declare the
former s16 and cast the latter to u16. Minimap_update 93.93->94.01,
fn_8013396C 96.78->97.40.

+++ src/main/dll/dll_0031_minimap.c
@@ -55,7 +55,7 @@ void fn_8013351C(void);
 
 
 extern void* Obj_GetPlayerObject(void);
-extern int Camera_GetViewportYOffset(void);
+extern s16 Camera_GetViewportYOffset(void);
 extern int objIsCurModelNotZero(int obj);
 extern void* gameTextGetBox(int box);
 extern void gameTextSetColor(u8 r, u8 g, u8 b, u8 a);
@@ -875,7 +875,7 @@ void fn_8013396C(void)
                 fn_80133818();
             }
             held = getButtonsHeld(0);
-            pressed = getButtonsJustPressed(0);
+            pressed = (u16)getButtonsJustPressed(0);
             if ((held & 0xc) == 0)
             {
                 if ((pressed & 1) != 0)
